+++ code/pyorg/scripts/filaments/oms_dst_analysis.py
@@ -46,12 +46,12 @@ rcParams['ytick.labelsize'] = 14
 ROOT_PATH = '/fs/pool/pool-ruben/antonio/filaments'
 
 # Input STAR files
-in_star = ROOT_PATH + '/ltomos_omsegs/omsegs_fil_ctrl/fil_ctrl_ltomos.star' # '/ltomos_omsegs/omsegs_1/omsegs_ltomos.star'
-in_wspace = ROOT_PATH + '/ana/oms_dsts/oms_fil_ctrl/oms_fil_ctrl_25_250_3_png_wspace.pkl' # None
+in_star = ROOT_PATH + '/ltomos_omsegs/omsegs_fil_new/fil_new_ltomos.star' # '/ltomos_omsegs/omsegs_fil_ctrl/fil_ctrl_ltomos.star' # '/ltomos_omsegs/omsegs_1/omsegs_ltomos.star'
+in_wspace = None # ROOT_PATH + '/ana/oms_dsts/oms_fil_ctrl/oms_fil_ctrl_25_250_3_png_wspace.pkl' # None
 
 # Output directory
-out_dir = ROOT_PATH + '/ana/oms_dsts/oms_fil_ctrl' # '/ana/oms_dsts/oms_vols/'
-out_stem = 'oms_fil_ctrl_25_250_3_png_2'
+out_dir = ROOT_PATH + '/ana/oms_dsts/oms_new' # '/ana/oms_dsts/oms_fil_ctrl' # '/ana/oms_dsts/oms_vols/'
+out_stem = 'oms_fil_ctrl_25_250_3_png' # 'oms_fil_ctrl_25_250_3_png_2'
 
 # Analysis variables
 pt_per = 95 # %