+++ code/pyorg/scripts/syn2/syn3/syn_coloc_analysis.py
@@ -52,27 +52,27 @@ ROOT_PATH = '/fs/pool/pool-lucic2/antonio/workspace/psd_an/ex/syn3/org'
 # Input STAR file
 in_star_1 = ROOT_PATH + '/pre/ltomos/ltomos_pre_pre_v6_chr_ref_xd5nm/pre_pre_v6_chr_ref_xd5nm_ltomos.star' # '/col/ltomos/ltomos_all_scol_xd5nm_test/all_scol_xd5nm_test_ltomos.star' # '/pre/ltomos/ltomos_all_pre_pre_ss7.31px/all_pre_pre_ltomos.star' # '/pre/ref_nomb_1_clean/ltomos_pre_premb_gather_mask/pre_premb_gather_mask_ltomos.star' # '/ves_40/ltomos_tether_lap/ves_40_cleft_premb_mask_lap_ltomos.star_ltomos.star' # '/ves_ap/ltomos_ves_ap_cleft/ves_ap_cleft_ltomos.star' # '/pre/ref_nomb_1_clean/ltomos_pre_premb_ABC/pre_premb_ABC_ltomos.star' # '/pre/ref_nomb_1_clean/ltomos_clst_flt_high_lap/clst_flt_high_lap_ltomos.star' # '/pre/ref_nomb_1_clean/ltomos_pre_premb_mask_lap/pre_premb_mask_lap_ltomos.star'
 key_1, l_id1 = 'pre', 0 # 4
-in_star_2 = ROOT_PATH + '/../../syn2/org/tether/ltomos/ltomos_all_pre_tether_ss14.62px/all_pre_tether_ltomos.star' # '/col/ltomos/ltomos_all_scol_xd5nm_test/all_scol_xd5nm_test_ltomos.star' # '/pre/ltomos/ltomos_all_pre_pre_ss7.31px/all_pre_pre_ltomos.star' # '/pre/ref_nomb_1_clean/ltomos_pre_premb_gather_mask/pre_premb_gather_mask_ltomos.star' # '/ves_40/ltomos_tether_lap/ves_40_cleft_premb_mask_lap_ltomos.star_ltomos.star' # '/ves_ap/ltomos_ves_ap_cleft/ves_ap_cleft_ltomos.star' # '/pre/ref_nomb_1_clean/ltomos_pre_premb_ABC/pre_premb_ABC_ltomos.star' # '/pre/ref_nomb_1_clean/ltomos_clst_flt_high_lap/clst_flt_high_lap_ltomos.star' # '/pre/ref_nomb_1_clean/ltomos_pre_premb_mask_lap/pre_premb_mask_lap_ltomos.star' # '/tether/ltomos/ltomos_all_pre_tether_ss7.31px/all_pre_tether_ltomos.star' # '/ves_40/ltomos_tether_lap/ves_40_cleft_premb_mask_lap_ltomos.star_ltomos.star' # '/pst/ref_nomb_1_clean/ltomos_pst_premb_gather_mask/pst_premb_gather_mask_ltomos.star' # '/pst/ampar_vs_nmdar/org/ltomos/ltomos_an_pre_2/an_pre2_ltomos.star' # '/ves_40/ltomos_tether_lap/ves_40_cleft_premb_mask_lap_ltomos.star_ltomos.star' # '/ves_ap/ltomos_ves_ap_cleft/ves_ap_cleft_ltomos.star' # '/az/ref_nomb_1_clean/ltomos_az_ref_1_gather_premb_cleft/az_cleft_premb_mask_ltomos.star' # '/pst/ampar_vs_nmdar/org/ltomos/ltomos_an_gather_pre_2/an_gather_pre2_ltomos.star' # '/pst/ampar_vs_nmdar/org/ltomos/ltomos_an_pre_2/an_pre2_ltomos.star' # '/pst/ref_nomb_1_clean/ltomos_pst_premb_gather_mask/pst_premb_gather_mask_ltomos.star' # '/ves_40/ltomos_tether_lap/ves_40_cleft_premb_mask_lap_ltomos.star_ltomos.star' # '/ves_40/ltomos_tether/ves_40_cleft_premb_mask_ltomos.star' # '/ves_40/ltomos_lap/lap_ltomos.star' # '/ves_40/ltomos_premb_mask/premb_mask_ltomos.star'
+in_star_2 = ROOT_PATH + '/tether/ltomos/ltomos_pre_tether_xd5nm/pre_tether_xd5nm_ltomos.star' # '/../../syn2/org/tether/ltomos/ltomos_all_pre_tether_ss14.62px/all_pre_tether_ltomos.star' # '/col/ltomos/ltomos_all_scol_xd5nm_test/all_scol_xd5nm_test_ltomos.star' # '/pre/ltomos/ltomos_all_pre_pre_ss7.31px/all_pre_pre_ltomos.star' # '/pre/ref_nomb_1_clean/ltomos_pre_premb_gather_mask/pre_premb_gather_mask_ltomos.star' # '/ves_40/ltomos_tether_lap/ves_40_cleft_premb_mask_lap_ltomos.star_ltomos.star' # '/ves_ap/ltomos_ves_ap_cleft/ves_ap_cleft_ltomos.star' # '/pre/ref_nomb_1_clean/ltomos_pre_premb_ABC/pre_premb_ABC_ltomos.star' # '/pre/ref_nomb_1_clean/ltomos_clst_flt_high_lap/clst_flt_high_lap_ltomos.star' # '/pre/ref_nomb_1_clean/ltomos_pre_premb_mask_lap/pre_premb_mask_lap_ltomos.star' # '/tether/ltomos/ltomos_all_pre_tether_ss7.31px/all_pre_tether_ltomos.star' # '/ves_40/ltomos_tether_lap/ves_40_cleft_premb_mask_lap_ltomos.star_ltomos.star' # '/pst/ref_nomb_1_clean/ltomos_pst_premb_gather_mask/pst_premb_gather_mask_ltomos.star' # '/pst/ampar_vs_nmdar/org/ltomos/ltomos_an_pre_2/an_pre2_ltomos.star' # '/ves_40/ltomos_tether_lap/ves_40_cleft_premb_mask_lap_ltomos.star_ltomos.star' # '/ves_ap/ltomos_ves_ap_cleft/ves_ap_cleft_ltomos.star' # '/az/ref_nomb_1_clean/ltomos_az_ref_1_gather_premb_cleft/az_cleft_premb_mask_ltomos.star' # '/pst/ampar_vs_nmdar/org/ltomos/ltomos_an_gather_pre_2/an_gather_pre2_ltomos.star' # '/pst/ampar_vs_nmdar/org/ltomos/ltomos_an_pre_2/an_pre2_ltomos.star' # '/pst/ref_nomb_1_clean/ltomos_pst_premb_gather_mask/pst_premb_gather_mask_ltomos.star' # '/ves_40/ltomos_tether_lap/ves_40_cleft_premb_mask_lap_ltomos.star_ltomos.star' # '/ves_40/ltomos_tether/ves_40_cleft_premb_mask_ltomos.star' # '/ves_40/ltomos_lap/lap_ltomos.star' # '/ves_40/ltomos_premb_mask/premb_mask_ltomos.star'
 key_2, l_id2 = 'tether', 1 # 1
 in_star_3 = ROOT_PATH + '/pst/ltomos/ltomos_pst_pre_v6_chr_ref_xd5nm/pst_pre_v6_chr_ref_xd5nm_ltomos.star' # '/col/ltomos/ltomos_all_scol_xd5nm_test/all_scol_xd5nm_test_ltomos.star' # '/pst/ltomos/ltomos_all_pst_pre_ss7.31px/all_pst_pre_ltomos.star' # '/pst/ref_nomb_1_clean/ltomos_pst_premb_gather_mask/pst_premb_gather_mask_ltomos.star' # '/pst/ampar_vs_nmdar/org/ltomos/ltomos_an_pre_2/an_pre2_ltomos.star' # '/pst/nrt/k2_ABC/ltomos_k2_premb_ABC/k2_premb_ABC_ltomos.star' # '/pst/ampar_vs_nmdar/org/ltomos/ltomos_an_gather_pre_2/an_gather_pre2_ltomos.star' # '/pst/ampar_vs_nmdar/org/ltomos/ltomos_ampar_nmdar_premb_mask/ampar_nmdar_premb_gather_mask_ltomos.star' # '/pst/nrt/ltomos_k4_gather_premb_mask/k4_gather_premb_mask_ltomos.star' # '/pst/nrt/ltomos_clst_flt_high_lap/clst_flt_high_lap_ltomos.star' # '/pst/nrt/ltomos_k2_premb_mask/k2_premb_mask_ltomos.star' # '/pst/nrt/ltomos_pst_premb_mask_lap/pst_premb_mask_lap_ltomos.star'
 key_3, l_id3 = 'pst', 0 # 4
-in_star_4 = ROOT_PATH + '/../../syn2/org/tether/ltomos/ltomos_all_pre_tether_ss14.62px/all_pre_tether_ltomos.star' # '/tether/ltomos/ltomos_all_pre_tether_ss7.31px/all_pre_tether_ltomos.star' # '/pst/ampar_vs_nmdar/org/ltomos/ltomos_ampar_nmdar_premb_mask/ampar_nmdar_premb_gather_mask_ltomos.star' # '/pst/nrt/ltomos_k4_gather_premb_mask/k4_gather_premb_mask_ltomos.star' # '/pst/nrt/k2_ABC/ltomos_k2_premb_ABC/k2_premb_ABC_ltomos.star' # '/pst/nrt/ltomos_clst_flt_high_lap/clst_flt_high_lap_ltomos.star' # '/pst/nrt/ltomos_k2_premb_mask/k2_premb_mask_ltomos.star' # '/pst/nrt/ltomos_pst_premb_mask_lap/pst_premb_mask_lap_ltomos.star'
+in_star_4 = ROOT_PATH + '/tether/ltomos/ltomos_pre_tether_xd5nm/pre_tether_xd5nm_ltomos.star' # '/../../syn2/org/tether/ltomos/ltomos_all_pre_tether_ss14.62px/all_pre_tether_ltomos.star' # '/tether/ltomos/ltomos_all_pre_tether_ss7.31px/all_pre_tether_ltomos.star' # '/pst/ampar_vs_nmdar/org/ltomos/ltomos_ampar_nmdar_premb_mask/ampar_nmdar_premb_gather_mask_ltomos.star' # '/pst/nrt/ltomos_k4_gather_premb_mask/k4_gather_premb_mask_ltomos.star' # '/pst/nrt/k2_ABC/ltomos_k2_premb_ABC/k2_premb_ABC_ltomos.star' # '/pst/nrt/ltomos_clst_flt_high_lap/clst_flt_high_lap_ltomos.star' # '/pst/nrt/ltomos_k2_premb_mask/k2_premb_mask_ltomos.star' # '/pst/nrt/ltomos_pst_premb_mask_lap/pst_premb_mask_lap_ltomos.star'
 key_4, l_id4 = 'sv', 3 # 1
 in_star_5 = ROOT_PATH + '/pst/ltomos/ltomos_pst_pre_v6_chr_ref_xd5nm/pst_pre_v6_chr_ref_xd5nm_ltomos.star'
 key_5, l_id5 = 'ampar', 0
 in_star_6 = ROOT_PATH + '/pst/ltomos/ltomos_pst_pre_v6_chr_ref_xd5nm/pst_pre_v6_chr_ref_xd5nm_ltomos.star'
 key_6, l_id6 = 'nmdar', 0
 in_tethers_csv = ROOT_PATH + '/../../syn/sub/relion/fils/pre/ref_nomb_1_clean/py_scripts/syn_num_tethers_2.csv'
-in_part = ROOT_PATH + '/../../syn/sub/relion/fils/pre/vtps/sph_rad_5_surf.vtp'
+in_part = ROOT_PATH + '/../../syn/sub/relion/fils/pre/vtps/sph_rad_7.31_surf.vtp'
 in_wspace = None # ROOT_PATH + '/pst/ampar_vs_nmdar/org/method_2/3_pre_tether_pst/aln_15_cr10/3_pre_tether_pst_sim20_wspace.pkl'
 in_tomo_res = ROOT_PATH + '/../rec/syn_scale_tomos.star' # None
 
 # Output directory
 out_dir = ROOT_PATH + '/coloc/3_pre_tether_pst/aln_15_cr30_xd5nm' # '/pst/ampar_vs_nmdar/org/col_scol/col_an_0_aln_clst_15_sim_200_nn1_1_nn2_1_nn3_1'
-out_stem = '3_pre_tether_pst_sim5_m1_v3'
+out_stem = '3_pre_tether_pst_sim200_m1'
 
 # Pre-processing variables
-pre_ssup = 5 #nm
+pre_ssup = 10 #nm
 pre_min_parts = 0
 
 # Analysis variables
@@ -84,7 +84,7 @@ ana_aln_dst = 15 # nm
 
 # P-value computation settings
 # Simulation model (currently only CSRV)
-p_nsims = 5
+p_nsims = 200
 p_per = 5 # %the
 
 # Figure saving options
@@ -490,12 +490,12 @@ if in_wspace is None:
             #     tomos_npc_l3_sim[tkey].append(.0)
                 # tomos_tca_sim[tkey].append(.0)
             sim_lyr_1 = gen_layer_model(ltomo_1, part_vtp, ModelCSRV, mode_emb='center')
-            sim_lyr_2 = gen_layer_model(ltomo_1, part_vtp, ModelCSRV, mode_emb='center')
+            # sim_lyr_3 = gen_layer_model(ltomo_3, part_vtp, ModelCSRV, mode_emb='center')
             # try:
-            # cfinder3 = ColumnsFinder(sim_lyr_1, ltomo_2, ltomo_3,
-            #                          ana_col_rad_v, ana_aln_dst_v, clst_method=ana_clst_method)
-            cfinder3 = ColumnsFinder(sim_lyr_1, sim_lyr_2, ltomo_3,
+            cfinder3 = ColumnsFinder(sim_lyr_1, ltomo_2, ltomo_3,
                                      ana_col_rad_v, ana_aln_dst_v, clst_method=ana_clst_method)
+            # cfinder3 = ColumnsFinder(sim_lyr_1, ltomo_2, sim_lyr_3,
+            #                          ana_col_rad_v, ana_aln_dst_v, clst_method=ana_clst_method)
             tomos_nc_sims2[tkey].append(cfinder3.get_num_columns())
             tomos_nsc_sims2[tkey].append(cfinder3.get_num_sub_columns())
             tomos_den_sims2[tkey].append(cfinder3.get_den_columns())